requirements.txt
streamlit==1.38.0
sentence-transformers==2.7.0
numpy==1.26.4
scikit-learn==1.5.1
python-dotenv==1.0.1
PyPDF2==3.0.1
groq==0.12.0
openai==1.35.13
torch==2.2.2
requests==2.32.3
tqdm==4.66.4
